raspberry to firebase edited but jeet has to do some modifications
